\name{agreementRateBinM}
\alias{agreementRateBinM}
\title{Agreement rate by maximum posterior values.}
\usage{agreementRateBinM(PjV1, PjV2)}
\description{Agreement rate by maximum posterior values.}
\value{agreement rate.}
\arguments{\item{PjV1}{Posterior matrix view 1 (by document).}
\item{PjV2}{Posterior matrix view 2 (by document).}}
