\name{assignIdxPerClMBinEM}
\alias{assignIdxPerClMBinEM}
\title{Assign final indices to data by maximum posterior value.}
\usage{assignIdxPerClMBinEM(PjV1, PjV2)}
\description{Assign final indices to data by maximum posterior value.}
\value{vector of indices for each data point.}
\arguments{\item{PjV1}{Posterior matrix view 1 (by document).}
\item{PjV2}{Posterior matrix view 2 (by document).}}
