\name{mApplyBern}
\alias{mApplyBern}
\title{Calculate Bernoulli likelihood for a rows of binary events...}
\usage{mApplyBern(X, prob)}
\description{Calculate Bernoulli likelihood for a rows of binary events}
\value{a matrix of Bernoulli likelihoods}
\arguments{\item{X}{a matrix of binary events (row-wise)}
\item{prob}{the Bernoulli probability}}
