\name{mApplyCat}
\alias{mApplyCat}
\title{Calculate categorical likelihood row-wise for categorical events...}
\usage{mApplyCat(X, prob)}
\description{Calculate categorical likelihood row-wise for categorical events}
\value{a matrix of categorical likelihoods}
\arguments{\item{X}{a matrix of categorical events (row-wise)}
\item{prob}{the categorical probability matrix (rows along events, cols along event values)}}
