\name{viewsClasses}
\alias{viewsClasses}
\title{Counts unique values in both views...}
\usage{viewsClasses(view1, view2)}
\description{Counts unique values in both views
Stops on any non-numeric values}
\value{list containing unique values for each view}
\arguments{\item{view1}{View 1}
\item{view2}{View 2}}
