\name{mApplyBern}
\alias{mApplyBern}
\title{Calculate Bernoulli likelihood row-wise for binary events...}
\usage{mApplyBern(X, prob)}
\description{Calculate Bernoulli likelihood row-wise for binary events}
\value{a matrix of Bernoulli likelihoods}
\arguments{\item{X}{a matrix of binary events (row-wise)}
\item{prob}{the Bernoulli probability (a single number)}}
